No longer making copies of structures for iteration. Speeds launch time considerably.
Added launch logging while loading state from disk.

Switched hash to be 64bit. Upper 32 bits is previous additive hash reduciton lower 32 bits is a 'merge 32 bit hashes into a 64 bit number and hash that' reduction. The addative would overlap if the same view blocks were found in any order, the new hash should be much more resilant.

No longer storing the input data or the decomosed view of that data in the agent. Leaving that up to the observer to store out in an appropriate format if desired. Currently not storing it.

adjusting somethingNew injection to actually respect the exploration factor, even if values are negative.

adjusting exploration factor down to 1, so there is roughly a 1:1 chance of taking a new action versus taking the best known action when a new action is found.

seperating the acting rate from the observing rate. Reduced acting rate to ~7.5 times a second, but still observing at ~30 times a second.

Added found rate to logging to show the rate of new states found during this session.

Adjusted the rate at which the lookAhead is adjusted. It now adjusts much more quickly to a stable value and only raises in ticks where the lookAhead was underutilized (but not when unsued to prevent it from trending upwards infiniately)

Guarded lookAhead to stay within 10-100,000 range to keep things sane and prevent rollover issues.